2018 03 14 17：55 画出个点坐标并连线

+++ app/src/main/java/com/cuipengyu/mbooktest/AppScreenUtil.java
@@ -14,11 +14,11 @@ public class AppScreenUtil {
         return mAppScreenUtil;
     }
 
-    public float getAppWidth() {
+    public int getAppWidth() {
         return AppConnextUtil.getAppConnect().getResources().getDisplayMetrics().widthPixels;
     }
 
-    public float getAppHeight() {
+    public int getAppHeight() {
         return AppConnextUtil.getAppConnect().getResources().getDisplayMetrics().heightPixels;
     }
 }
